completed folder restructure
