+++ src/graphics/basicEffect.mjs
@@ -1,4 +1,4 @@
-import shaders from "../shaders.mjs";
+import shaders from "./shaders.mjs";
 import Effect from "./effect.mjs";
 
 export default class BasicEffect extends Effect {
