+++ src/susurrus.mjs
@@ -1,10 +1,22 @@
-export { default as Color } from "./color.mjs";
-export { default as Colors } from "./colors.mjs";
 export { default as Sketch } from "./sketch.mjs";
-export { default as Polygon } from "./polygon.mjs";
-export { default as PolygonData } from "./polygonData.mjs";
-export { default as Effect } from "./graphics/effect.mjs";
+
+export { default as AttributeElement } from "./graphics/attributeElement.mjs";
+export { default as AttributeSchema } from "./graphics/attributeSchema.mjs";
+export { default as AttributeType } from "./graphics/attributeType.mjs";
 export { default as BasicEffect } from "./graphics/basicEffect.mjs";
-export { default as Camera } from "./camera.mjs";
+export { default as Color } from "./graphics/color.mjs";
+export { default as Colors } from "./graphics/colors.mjs";
+export { default as DrawMode } from "./graphics/drawMode.mjs";
+export { default as Effect } from "./graphics/effect.mjs";
+export { default as IndexBuffer } from "./graphics/indexBuffer.mjs";
+export { default as Polygon } from "./graphics/polygon.mjs";
+export { default as PolygonData } from "./graphics/polygonData.mjs";
+export { default as VertexBuffer } from "./graphics/vertexBuffer.mjs";
+export { default as VertexUsage } from "./graphics/vertexUsage.mjs";
+
 export { default as Matrix } from "./maths/matrix.mjs";
+export { default as Vector2 } from "./maths/vector3.mjs";
 export { default as Vector3 } from "./maths/vector3.mjs";
+export { default as Vector4 } from "./maths/vector3.mjs";
+
+export { default as Camera } from "./utilities/camera.mjs";
